Refactor the modals to components.
